# Port checker to WASM