--- a/.gitignore
+++ b/.gitignore
@@ -20,3 +20,7 @@ node_modules
 
 # how does this get here
 doc/VERSION
+
+*.wasm
+
+target