--- a/Makefile
+++ b/Makefile
@@ -1,7 +1,9 @@
 NODE_MODULES = node_modules
 VERSION := $(shell cat ./VERSION)
 
-.PHONY: build assets deps lint prebaked-build test
+export RUSTFLAGS=-Ctarget-feature=+simd128
+
+.PHONY: build assets deps lint prebaked-build test wasm
 
 assets:
 	npm run assets
@@ -25,3 +27,7 @@ prebaked-build:
 
 test:
 	npm run test
+
+wasm:
+	cargo build --release --target wasm32-unknown-unknown
+	cp -vf ./target/wasm32-unknown-unknown/release/*.wasm ./web/static/wasm