--- a/wasm/anubis/src/lib.rs
+++ b/wasm/anubis/src/lib.rs
@@ -1,4 +1,3 @@
-use std::boxed::Box;
 use std::sync::{LazyLock, Mutex};
 
 #[cfg(target_arch = "wasm32")]