--- a/wasm/anubis/src/lib.rs
+++ b/wasm/anubis/src/lib.rs
@@ -1,5 +1,10 @@
 use std::sync::{LazyLock, Mutex};
 
+extern crate wee_alloc;
+
+#[global_allocator]
+static ALLOC: wee_alloc::WeeAlloc = wee_alloc::WeeAlloc::INIT;
+
 #[cfg(target_arch = "wasm32")]
 mod hostimport {
     use crate::{DATA_BUFFER, DATA_LENGTH};