--- a/web/static/wasm/.gitignore
+++ b/web/static/wasm/.gitignore
@@ -1,0 +1,2 @@
+*
+!.gitignore