This is for the pivot strategy to accomplish #94. Overall the strategy looks like this:

1. Port existing checker to WebAssembly
2. Hook it into the existing flow/logic
3. Switch default checker to use this new webassembly-based flow
4. Implement a small set of other checkers (equi-x, etc) and randomly pick between them based on bytes in the challenge.

NOTE: this changes the proof of work calculation logic for the new checker.

This differs from the JavaScript implementations by constructing the hash differently. In JavaScript implementations, the SHA-256 input is the result of appending the nonce as an integer to the hex-formatted challenge, eg:

```js
sha256(`${challenge}${nonce}`);
```

This **does work**, however I think that this can be done a bit better by operating on the challenge bytes _directly_ and treating the nonce as a salt.

The nonce is also randomly encoded in either big or little endian depending on the last byte of the data buffer in an effort to make it more annoying to automate with GPUs.