@max-te 2025-04-12
Hi, so I have implemented something like this with Rust and WASM before for proof-of-work on login requests. That code lives in my private repo, but I've contributed some work back to the sha2 crate: If you want to use the SHA2 family of hashes, I recommend you check out the 0.11 pre-release version of the sha2 crate and compile the wasm with `RUSTFLAG=-C target-feature=+simd128`. This activates the [wasm32 simd backend](https://github.com/RustCrypto/hashes/pull/562) which gives around an 8x speed-up..