@Xe 2025-04-12
I'm gong to have to see if wazero supports SIMD, if it does then this will be an obvious addition. Thanks!