@Xe 2025-04-12
```
go tool benchcmp before-simd.txt after-simd.txt 
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark              old ns/op     new ns/op     delta
BenchmarkSHA256-32     13222850      8209668       -37.91%

benchmark              old allocs     new allocs     delta
BenchmarkSHA256-32     85             85             +0.00%

benchmark              old bytes     new bytes     delta
BenchmarkSHA256-32     29838         23071         -22.68%
```
very yes, neat!